Version 0.1.2\n

+++ deploy.sh
@@ -8,18 +8,19 @@ if [ -z ${VERSION} ]; then
 fi
 
 git status
-
 echo -e "Continue? (Ctrl+C to cancel)"
-read
+read CONTINUE
 
-git add .
-git commit -m "Version ${VERSION}\n$2" && \
-    git push github master --tags ${VERSION}
+if [[ "y" == ${CONTINUE} ]] || [[  "Y" == ${CONTINUE} ]]; then
+    git add .
+    git commit -m "Version ${VERSION}\n$2" && \
+        git push github master --tags ${VERSION}
 
-sudo docker build . --force-rm --rm -t chrisdlangton/h2o-flow:${VERSION} --compress && \
-    sudo docker push chrisdlangton/h2o-flow:${VERSION} && \
-    sudo docker push chrisdlangton/h2o-flow:latest && \
-    echo ok && exit
+    sudo docker build . --force-rm --rm -t chrisdlangton/h2o-flow:${VERSION} --compress && \
+        sudo docker push chrisdlangton/h2o-flow:${VERSION} && \
+        sudo docker push chrisdlangton/h2o-flow:latest && \
+        echo ok && exit
 
-echo Failed
-exit 1
+    echo Failed
+    exit 1
+fi
